removed docking from vraytoolbox due to crashing bug when a new scene is created
